minor changes in the source tree
